producer/requirements.txt
psycopg2-binary
confluent-kafka
python-dotenv
schedule